+++ CHANGELOG.md
@@ -2,7 +2,17 @@
 
 ## Unreleased
 
-- Published `generator` module
+### Changed
+
+- Internal random number generator interface for `V7Generator` from
+  `rand::RngCore` to `generator::Rng` to relax hard dependency on `rand` crate
+  - Maintained compatibility by implementing `generator::Rng` for
+    `rand::RngCore`
+
+### Added
+
+- `V7Generator::with_rand08()` to make integration with `rand` explicit
+- `generator` module
 
 ## v1.0.2 - 2024-09-04